% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_aaa_main_lib.R
\name{get_match_set}
\alias{get_match_set}
\title{Get the subset of a Population of Classifiers that matches a given State}
\usage{
get_match_set(instance_state, pop)
}
\arguments{
\item{instance_state}{A state from the RLCS environment}

\item{pop}{A population of Classifiers}
}
\value{
Numeric vector, of indices of the matching Classifiers
}
\description{
Get the subset of a Population of Classifiers that matches a given State
}
\examples{
demo_env1 <- rlcs_demo_secret1()
demo_params <- RLCS_hyperparameters(n_epochs = 280, deletion_trigger = 40, deletion_threshold = 0.9)
rlcs_model1 <- rlcs_train_sl(demo_env1, demo_params)
print(rlcs_model1)
get_match_set("00101", rlcs_model1)

}
