% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnist_bin01_49b}
\alias{mnist_bin01_49b}
\title{Compressed images for 0 and 1 - Sample from MNIST}
\format{
## `mnist_bin01_49b`
A data frame with 12665 rows and 3 columns:
\describe{
  \item{sharp_image}{28*28 image but compressed into BINARY format}
  \item{state}{7*7 image further compressed binary format}
  \item{class}{class of the image, 0 or 1 only}
}
}
\source{
Famous MNIST data set by Yann Lecun
}
\usage{
mnist_bin01_49b
}
\description{
Compressed images for 0 and 1 - Sample from MNIST
}
\keyword{datasets}
