% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fits}
\alias{plot_fits}
\title{plot_fits}
\usage{
plot_fits(dataframe = NA, par = "par")
}
\arguments{
\item{dataframe}{dataframe object containing the data}

\item{par}{character string, dataframe column name containing par levels}
}
\value{
plot showing all fits contained in the dataframe (+/- upper and lower CI)
}
\description{
Plot multiple e and p fits with u/l bounds on one plot
}
