% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{plot_params}
\usage{
plot_params(dataframe = NA, treatment = NA)
}
\arguments{
\item{dataframe}{av_params dataframe derived from fit_rlcs}

\item{treatment}{character string, dataframe column name containing treatment levels}
}
\value{
4 panel plot showing average Fv/Fm, rETRmax, alpha and Ek (+/- SE) per treatment
}
\description{
Plot RLC average parameters (fvfm, rETRmax, alpha, ek) +/- SE after calculation with rlc_fit
}
