% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rlcs.R
\name{plot_rlcs}
\alias{plot_rlcs}
\title{plot_rlcs}
\usage{
plot_rlcs(dataframe = NA, par = "par", parameter = "etr",
  treatment = NA, error = NA)
}
\arguments{
\item{dataframe}{dataframe object containing the data}

\item{par}{character string, dataframe column name containing par levels}

\item{parameter}{character string, dataframe column name containing desired parameter data, e.g. etr, npq etc.}

\item{treatment}{character string, dataframe column name containing treatment levels}

\item{error}{character string, dataframe column name containing SE data}
}
\value{
plot showing all treatments parameter (+/- SE)
}
\description{
Plot multiple rlcs on one plot
}
