% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlc_fit.R
\name{rlc_fit}
\alias{rlc_fit}
\title{rlc_fit}
\usage{
rlc_fit(dataframe = NA, par = "par", yield = "y", f = "f",
  fm. = "fm.", sample = NA, treatment = NA, predict_ul = T)
}
\arguments{
\item{dataframe}{name of the dataframe containing the data}

\item{par}{character string of dataframe column containing PAM par levels}

\item{yield}{character string, dataframe column containing Y(PSII) data}

\item{f}{character string, dataframe column containing F data}

\item{fm.}{character string, dataframe column containing Fm' data}

\item{sample}{character string, dataframe column containing sample names per light curve}

\item{treatment}{character string, dataframe column containing treatment levels if applicable. If provided, averages are made per level of treatment for variables (rETR, NPQ, YNPQ, YNO) and parameters (ETRmax, alpha, Ek, Fv/Fm)}

\item{predict_ul}{verbose (T/F), provide predictions of fit, upper and lower bounds for e and p models}
}
\value{
rlc_results list containing the following dataframes:
   1. 'rlcs' containing the original data, with calculate variables (rETR, NPQ, YNPQ, YNO)
      and e and p model parameters (a,b,c) and derived parameters (rETRmax, alpha, Ek)

   If predict_ul ==T:
   2. 'fits' If predict_ul == Tcontaining predicted fit, upper and lower bounds for each
      light curve e and p model calculated using predictNLS function

   If 'treatment' is provided:
   3. 'av_variables' containing averages of rETR, NPQ, YNPQ and YNO per treatment,
       including standard deviation and standard error
   4. 'av_params' containing averages of rETRmax, alpha, ek and fvfm per treatment,
       including standard deviation and standard error
   5. 'av_fits' containing e and p fit of etr ~ par with upper and lower bounds per
       treatment.
}
\description{
simple fit of e and p model to PAM data.
}
