% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_datasets_datamining.R
\name{rlcs_demo_secret1}
\alias{rlcs_demo_secret1}
\title{A very simple demo dataset generator.}
\usage{
rlcs_demo_secret1()
}
\value{
a data frame with a "state" and a "class", 32 entries
}
\description{
A very simple demo dataset generator.
}
\examples{
demo_dataset1 <- rlcs_demo_secret1()
}
