% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_hyperparams_class.R
\docType{class}
\name{RLCS_hyperparameters-class}
\alias{RLCS_hyperparameters-class}
\alias{RLCS_hyperparameters}
\title{Generates an object of Hyperparameters required to run RLCS training.
All parameters have a default, although it is recommended each parameter is
reviewed and set depending on the use-case}
\value{
An RLCS Object with a set of hyperparameters for LCS Training.
}
\description{
Generates an object of Hyperparameters required to run RLCS training.
All parameters have a default, although it is recommended each parameter is
reviewed and set depending on the use-case
}
\section{Fields}{

\describe{
\item{\code{wildcard_prob}}{numeric. Probability of a matching rule to set wildcards
instead of fixed bits.}

\item{\code{rd_trigger}}{numeric. When to trigger Genetic Algorithm.
The smaller this number, the more often new rules are generated}

\item{\code{mutation_probability}}{numeric. For each new rule generated, what's the
probability of each bit to be mutated.}

\item{\code{parents_selection_mode}}{character.}

\item{\code{tournament_pressure}}{numeric.}

\item{\code{n_epochs}}{numeric.}

\item{\code{deletion_trigger}}{numeric.}

\item{\code{deletion_threshold}}{numeric.}
}}


\examples{
default_rlcs_hyperparameters <- RLCS_hyperparameters()
}
