% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_datasets_datamining.R
\name{rlcs_mux11}
\alias{rlcs_mux11}
\title{11-bits multiplexer as example dataset}
\usage{
rlcs_mux11()
}
\value{
A multiplexer dataset, with states (11-bits strings) and a class.
}
\description{
This example show-cases Epistasis. A genetic concept, points
to interactions between features.
}
