% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_datasets_datamining.R
\name{rlcs_mux6}
\alias{rlcs_mux6}
\title{6-bits multiplexer as example dataset}
\usage{
rlcs_mux6()
}
\value{
A multiplexer dataset, with states (6-bits strings) and a class.
}
\description{
This example show-cases Epistasis. A genetic concept, points
to interactions between features.
}
