% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_datasets_datamining.R
\name{rlcs_not_bit_4_10}
\alias{rlcs_not_bit_4_10}
\title{10 bits as example dataset}
\usage{
rlcs_not_bit_4_10()
}
\value{
A demo dataset with bit 4 negated as predicted class.
}
\description{
This example show-cases longer chains, simple info.
}
