% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_other_predict.R
\name{rlcs_predict_rl}
\alias{rlcs_predict_rl}
\title{Predict an Action for a given input set of states}
\usage{
rlcs_predict_rl(
  lcs,
  match_set,
  verbose = F,
  possible_actions = c("left", "right", "up", "down")
)
}
\arguments{
\item{lcs}{A trained RLCS model, with a population of classifiers.}

\item{match_set}{positions for matching rules. In RL, we focus on this to limit processing times.}

\item{verbose}{Detail or not the results? Defaults to FALSE.}

\item{possible_actions}{List of acceptable actions in the given "world".}
}
\value{
A vector of values of actions.
}
\description{
Predict an Action for a given input set of states
}
