% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_other_predict.R
\name{rlcs_predict_sl}
\alias{rlcs_predict_sl}
\title{Predict a Class for a given input set of states}
\usage{
rlcs_predict_sl(test_env_df, pop, verbose = F)
}
\arguments{
\item{test_env_df}{The dataset. Must contain a compatible state column.}

\item{pop}{A trained RLCS model, consisting of a population of classifiers.}

\item{verbose}{Detail or not the results? Defaults to FALSE.}
}
\value{
A vector of values of classes.
}
\description{
Predict a Class for a given input set of states
}
