% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_stone_lib.R
\name{rlcs_rosetta_decode_rule}
\alias{rlcs_rosetta_decode_rule}
\title{To facilitate reading a rule in a decoded format.}
\usage{
rlcs_rosetta_decode_rule(rule, rosetta_stone_obj)
}
\arguments{
\item{rule}{A rule from an RLCS model}

\item{rosetta_stone_obj}{A list previously generated by rlcs_rosetta_stone()}
}
\description{
Rosetta Stone encodes numerical data into binary strings for compatibility
with RLCS. This helper function helps decode the generated model.
}
