% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_stone_lib.R
\name{rlcs_rosetta_stone}
\alias{rlcs_rosetta_stone}
\title{Transforms numerical columns of a dataframe into RLCS compatible binary strings}
\usage{
rlcs_rosetta_stone(input_df, class_col = 1)
}
\arguments{
\item{input_df}{A dataframe with numeric or integers columns, and a "class" column}

\item{class_col}{The "class" column which is not to be transformed}
}
\value{
a list with a vector of cut values per column, column names and assigned values per entry, per column
}
\description{
Transforms Numeric Columns each into 16 buckets, binary-encoded with Gray encoding.
THIS IS CURRENTLY Work In Progress and should not be used aside from the Iris example.
}
\examples{
rlcs_iris <- rlcs_rosetta_stone(iris, class_col=5) ## NOT part of the LCS Algorithm!
head(rlcs_iris$model, n=3)
}
