% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_train_SL_specific.R
\name{rlcs_train_sl_horizontal_split}
\alias{rlcs_train_sl_horizontal_split}
\title{Train a Learning Classifier System (LCS), but go faster by running
and merging several agents in parallel on smaller environment subsets.
EXPERIMENTAL.}
\usage{
rlcs_train_sl_horizontal_split(
  train_env_df,
  run_params = RLCS_hyperparameters(),
  pre_trained_lcs = NULL,
  n_agents = 2,
  split_horizontal = T,
  max_pop_size_parallel = 10000,
  use_gpu = F
)
}
\arguments{
\item{train_env_df}{A data frame containing, specifically, one "state" and one
"class" column. The "state" column MUST contain strings made of ONLY 0 and 1,
such as: "00110101". This is a requirement for the current RLCS implementation.}

\item{run_params}{An RLCS_hyperparameters object, for which an object construction is provided.}

\item{pre_trained_lcs}{Optional. Can be used to EVOLVE a pre-trained LCS.}

\item{n_agents}{Default is 0. OPTIONAL. ONLY USED if foreach and doParallel are available.
IF available, a number of parallel cores, as indicated PRIOR to calling RLCS
like so: makeCluster() registerDoParallel()
then RLCS will train n_agents in parallel.}

\item{split_horizontal}{Defaults to T. If foreach and doParallel are available: Splits evenly across
N agents (N number of cores/threads) the input dataset.
Then trains N agents, and then merges the resulting data.
This can potentially speed-up the process, but will probably over-fit for each
subset, hence probably reducing overall model accuracy.}

\item{max_pop_size_parallel}{Defaults to 10000. OPTIONAL. ONLY USED if foreach and doParallel are available
Applies as last step: additional deletion to contain population sizes after merging.}

\item{use_gpu}{Defaults to FALSE. OPTIONAL. ONLY USED if torch is available. Defaults to CUDA if found.
Otherwise goes back to CPU (and is slower than not enabling the GPU option in the first place).}
}
\value{
An \R \code{RLCS Model} containing the proposed model, made of several classifiers.
}
\description{
Train a Learning Classifier System (LCS), but go faster by running
and merging several agents in parallel on smaller environment subsets.
EXPERIMENTAL.
}
\examples{
## Generate running hyperparameters
demo_params <- RLCS_hyperparameters(n_epochs = 400, deletion_trigger = 40, deletion_threshold = 0.9)
## One demo dataset for data mining scenario
demo_env1 <- rlcs_demo_secret1()
## Try to see for yourself what the dataset hides:
demo_env1
## Generate the model with RLCS:
rlcs_model <- rlcs_train_sl(demo_env1, demo_params)
print(rlcs_model)
plot(rlcs_model)
}
