% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_train_SL_specific.R
\name{rlcs_train_sl_parallel_search_space}
\alias{rlcs_train_sl_parallel_search_space}
\title{Train a Learning Classifier System (LCS). But try to cover more search space
by running and merging several agents in parallel, for the same iterations.
EXPERIMENTAL.}
\usage{
rlcs_train_sl_parallel_search_space(
  train_env_df,
  run_params = RLCS_hyperparameters(),
  pre_trained_lcs = NULL,
  n_agents = 2,
  use_validation = F,
  merge_best_n = 0,
  second_evolution_iterations = 1,
  second_evolution_run_params = NULL,
  max_pop_size_parallel = 10000,
  use_gpu = F
)
}
\arguments{
\item{train_env_df}{A data frame containing, specifically, one "state" and one
"class" column. The "state" column MUST contain strings made of ONLY 0 and 1,
such as: "00110101". This is a requirement for the current RLCS implementation.}

\item{run_params}{An RLCS_hyperparameters object, for which an object construction is provided.}

\item{pre_trained_lcs}{Optional. Can be used to EVOLVE a pre-trained LCS.}

\item{n_agents}{Default is 2. If foreach and doParallel are available, a number of parallel
cores, as indicated PRIOR to calling RLCS like so:
makeCluster() registerDoParallel()
then RLCS will train n_agents in parallel.}

\item{use_validation}{Default is FALSE. OPTIONAL. ONLY USED if foreach and doParallel are available.
When training several models in parallel, this parameter modifies selection of best
one by running a test in a validation set of 10 percent of samples, which is
first removed from the training set.}

\item{merge_best_n}{Default is 0. OPTIONAL.ONLY USED if foreach and doParallel are available.
Choose to merge and compact the best n (1 < n < n_agents) of your parallelly trained
agents. This includes a compaction previous to returning results, but will most
probably return a larger population as a trade-off for expecting better accuracy.}

\item{second_evolution_iterations}{Defaults to 1. OPTIONAL.ONLY USED if foreach and doParallel are available and
used. On top of the above, it will run a second "era", whereby only best agents
are surviving and competing again, only to then be merged.}

\item{second_evolution_run_params}{Defaults to NULL. OPTIONAL. ONLY USED if second_evolution_iterations is bigger
than 1. The idea here is that after a slower, more exploratory first era, a
second (and more) era(s) can be used to push more generalization.}

\item{max_pop_size_parallel}{Defaults to 10000. OPTIONAL. ONLY USED if foreach and doParallel are available
Applies as last step: additional deletion to contain population sizes after merging.}

\item{use_gpu}{Defaults to FALSE. OPTIONAL. ONLY USED if torch is available. Defaults to CUDA if found.
Otherwise goes back to CPU (and is slower than not enabling the GPU option in the first place).}
}
\value{
An \R \code{RLCS Model} containing the proposed model, made of several classifiers.
}
\description{
Train a Learning Classifier System (LCS). But try to cover more search space
by running and merging several agents in parallel, for the same iterations.
EXPERIMENTAL.
}
\examples{
## Generate running hyperparameters
demo_params <- RLCS_hyperparameters(n_epochs = 400, deletion_trigger = 40, deletion_threshold = 0.9)
## One demo dataset for data mining scenario
demo_env1 <- rlcs_demo_secret1()
## Try to see for yourself what the dataset hides:
demo_env1
## Generate the model with RLCS:
rlcs_model <- rlcs_train_sl(demo_env1, demo_params)
print(rlcs_model)
plot(rlcs_model)
}
