% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_aaa_main_lib.R
\name{reverse_match_set}
\alias{reverse_match_set}
\title{Returns all rlcs_environment entries (data frame row indices) that match a Classifier/Rule}
\usage{
reverse_match_set(rlcs_classifier, rlcs_environment)
}
\arguments{
\item{rlcs_classifier}{A Population of Classifiers, pre-trained.}

\item{rlcs_environment}{An RLCS environment (data frame inc. a "state" variable)}
}
\value{
Data Frame Row Indices from an environment that match a Classifier
}
\description{
Returns all rlcs_environment entries (data frame row indices) that match a Classifier/Rule
}
\examples{
## See Iris Example in GitHub for detailed example. NOT RUN
}
