% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_other_predict.R
\name{rlcs_predict_rl}
\alias{rlcs_predict_rl}
\title{Predict an Action for a given input set of states}
\usage{
rlcs_predict_rl(pop, verbose = F)
}
\arguments{
\item{pop}{A trained RLCS model population, consisting of a population of classifiers.}

\item{verbose}{Detail or not the results? Defaults to FALSE.}
}
\value{
A vector of values of actions.
}
\description{
Predict an Action for a given input set of states
}
