% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_train_SL_specific.R
\name{rlcs_simplify_pop}
\alias{rlcs_simplify_pop}
\title{Try to simplify a Learning Classifier System (LCS) population.}
\usage{
rlcs_simplify_pop(rlcs_obj, train_df)
}
\arguments{
\item{rlcs_obj}{An RLCS model object.}

\item{train_df}{The dataset used for training the model. This COULD be a different dataset,
although results might vary then.}
}
\value{
An \R \code{RLCS Model} containing a shorter population of rules, ideally.
}
\description{
By iterating over the rules of a population
assumed sorted already, going backwards, try to evaluate
the quality of the model and make sure you can maintain results while removing
rules one by one. Only remove rules that do not affect results.
}
\examples{
## Supposing you have trained an RLCS model against the iris dataset as
## proposed in the documented examples:
rlcs_iris <- rlcs_rosetta_stone(iris, class_col=5)
full_dataset <- cbind(iris, rlcs_iris$model)
full_dataset <- full_dataset[sample(1:nrow(full_dataset), nrow(full_dataset), replace = FALSE), ]
## Train-test separation:
train_set <- sample(1:nrow(full_dataset), size = round(0.8*nrow(full_dataset)), replace = FALSE)
train_environment <- full_dataset[train_set,]
test_environment <- full_dataset[-train_set,]
iris_classifier <- rlcs_train_sl(train_environment)
cleaner_iris_classifier <- rlcs_simplify_pop(iris_classifier, train_environment)
print(cleaner_iris_classifier)
print(cleaner_iris_classifier)
plot(cleaner_iris_classifier)
}
