% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLCS_train_SL_specific.R
\name{rlcs_train_sl}
\alias{rlcs_train_sl}
\title{Train a Learning Classifier System (LCS).}
\usage{
rlcs_train_sl(
  train_env_df,
  run_params = RLCS_hyperparameters(),
  pre_trained_lcs = NULL,
  verbose = FALSE,
  n_agents = 0,
  use_validation = F
)
}
\arguments{
\item{train_env_df}{A data frame containing, specifically, one "state" and one
"class" column. The "state" column MUST contain strings made of ONLY 0 and 1,
such as: "00110101". This is a requirement for the current RLCS implementation.}

\item{run_params}{An RLCS_hyperparameters object, for which an object construction is provided.}

\item{pre_trained_lcs}{Optional. Can be used to EVOLVE a pre-trained LCS.}

\item{verbose}{Default is FALSE. I strongly suggest you do not use that parameter.}

\item{n_agents}{Default is 0. OPTIONAL. ONLY USED if foreach and doParallel are available.
IF available, a number of parallel cores, as indicated PRIOR to calling RLCS
like so: makeCluster() registerDoParallel()
then RLCS will train n_agents parallel model and select the best one, by accuracy.}

\item{use_validation}{Default is FALSE. OPTIONAL.ONLY USED if foreach and doParallel are available.
When training several models in parallel, this parameter modifies selection of best
one by running a test in a validation set of 10 percent of samples, which is
first removed from the training set.}
}
\value{
An \R \code{RLCS Model} containing the proposed model, made of several classifiers.
}
\description{
Train a Learning Classifier System (LCS).
}
\examples{
## Generate running hyperparameters
demo_params <- RLCS_hyperparameters(n_epochs = 400, deletion_trigger = 40, deletion_threshold = 0.9)
## One demo dataset for data mining scenario
demo_env1 <- rlcs_demo_secret1()
## Try to see for yourself what the dataset hides:
demo_env1
## Generate the model with RLCS:
rlcs_model <- rlcs_train_sl(demo_env1, demo_params)
print(rlcs_model)
plot(rlcs_model)
}
